Extract enterprise response envelope module
